Create Week 4: Assignment 4 - Question 2
Given two arrays of positive integers, write a C Program to output the smallest number in the first array that is also present in the second one.

Input

The first line contains the size of the first array.
The second line contains the contents of first array. 
The third line contains the size of the second array.
The fourth line contains the contents of second array.