Create Week 4: Assignment 4 - Question 3
Given two strings(character arrays), write a C Program to check if one of them is an anagram of the other.

An anagram is a word or phrase formed by rearranging the letters of a different word or phrase, using all the original letters exactly once.

Examples are LISTEN and SILENT , KNEE and KEEN.

Input

The first line contains the size of the character array.
The second line contains the contents of first array. 
The third line contains the size of the second array.

Note: The maximum size of the character array can be assumed to be 20.
The contents of both arrays are only upper case alphabets.