Create Week 6: Assignment 6 - Question 1
Given an n×n integer Matrix A and an positive number ℓ such that 2ℓ+1≤n, write a C program to print the ℓ window smoothing of A. 

To get the  ℓ-window smoothing of A , we replace A[i][j] with the sum of the values of the imaginary submatrix S of A with centre at A[i][j], and having size 2ℓ+1×2ℓ+1


More precisely, the smoothed matrix B[i,j]=∑u=ilih∑v=jljhA[u][v]  
where il=max(i−ℓ,0),ih=min(i+ℓ,n−1),jl=max(j−ℓ,0),jh=min(j+ℓ,n−1).

Input
The first line contains the dimension of the matrix n. Assume n < 100.
The second line contains the smoothing parameter ℓ. 
The next n lines contains the contents of the matrix A, each row per line.

Output
The smoothed matrix of A

Note: Ignore the Passed after ignoring Presentation Error comment.